improve: security, code quality, and reliability fixes
- Add config.Validate() to enforce required JWT/AES secrets at startup
- Add rate limiting middleware (10 req/min login, 100 req/min on-login)
- Add MIME magic bytes validation for logo uploads (PNG/JPEG/WebP)
- Add GET /api/v1/health endpoint (Postgres, Redis, Engine status)
- Add CRUD endpoints for profile price mappings (List/Update/Delete)
- Add audit logging infrastructure (repo, middleware, auth handler)
- Refactor ExportCSV/Excel to share fetchSalesForRange() helper
- Fix worker cache keys to use appcache helper functions
- Fix voucherSessionCleanup to actually scan and delete expired Redis keys
- Add Ping() and Client() methods to Cache
- Fix all integration test circular dependencies and auth test failures

+++ internal/roskit/pipeline/timeseries/influx_integration_test.go
@@ -137,7 +137,7 @@ func TestInfluxReader_QueryRange_ReturnsData(t *testing.T) {
 	require.NoError(t, w.WritePoint(ctx, timeseries.Point{
 		Measurement: "system_resource",
 		Tags:        map[string]string{"router_id": routerID},
-		Fields:      map[string]any{"cpu-load": float64(42)},
+		Fields:      map[string]any{"cpu-load": int64(42)},
 		Timestamp:   ts,
 	}))
 	require.NoError(t, w.Flush(ctx))
